Move extract to scripts
