+++ com.ibm.streamsx.topology/bin/spl-python-extract.py
@@ -1,7 +1,7 @@
 # Licensed Materials - Property of IBM
 # Copyright IBM Corp. 2016,2017
 import sys
-import streamsx.spl.extract as sse
+import streamsx.scripts.extract as sse
 
 sse._extract_from_toolkit()
 sys.exit(0)
