+++ com.ibm.streamsx.topology/opt/python/packages/streamsx/scripts/extract.py
